Add possibility to ignore files with .rgignore in the project dir
A new function was created for find_files.
Was lucky enough to have find_files accept the find command.

A different approach was needed for grep where a setting on the level of
telescope needed to be set (`vimgrep_arguments`).

As a result of this there is some inconsistency where some keybindings
are done with pure lua and others are viml.

Also, passing the theme to `find_files` while it has parameters means
first passing the parameters to the theme and then just passing the
theme to the picker.